# specs/007-integration-gateway-and-mcp/contracts/agent-session.http-openapi.yaml
openapi: 3.0.3
info:
  title: agent_session.host.title
  version: 1.0.0
  x-title-i18n: {zh-CN: 插件服务会话, en-US: Plugin service sessions}
servers:
  - url: /
security:
  - STS: []
x-contract:
  implementation_status: current_baseline
  target_runtime_contract:
    spec: specs/031-agent-runtime-durable-scheduling/spec.md
    schema: specs/031-agent-runtime-durable-scheduling/contracts/run-state.openapi.yaml
    status: proposed_not_implemented
    invocation_identity: one_message_one_run
    process_restart: resume_or_explicit_unrecoverable_after_reconciliation
    event_resume: run_id_and_event_seq
    sse_role: subscription_only
    note: Existing invocation schemas and x-contract fields below describe the current wire contract; update them with contract tests during 031 T002/T009 migration.
  execution_context:
    history_source: core_persisted_messages_before_selected_message
    snapshot: captured_under_session_lock_before_invocation
    maximum_history_messages: 128
    maximum_history_content_bytes: 262144
    history_overflow: AGENT_SESSION_CONFLICT
    pending_state_source: core_runtime_structured_task
    pending_state_ttl_seconds: 259200
    pending_state_expired: AGENT_SESSION_CONTEXT_EXPIRED
    commit: atomic_with_successful_invocation_and_assistant_message
    client_state_override: forbidden
  actor: sts_plugin_service_actor
  api_key_supported: false
  owner_scope: [tenant_uuid, plugin_id, service_actor]
  identity_override: reject
  external_message_roles: [user]
  idempotency:
    header: Idempotency-Key
    scope: [session_uuid, operation, key]
    ttl_seconds: 86400
    expired_key: AGENT_SESSION_IDEMPOTENCY_EXPIRED
    changed_request: AGENT_SESSION_CONFLICT
    retry_execution_automatically: false
  invocation:
    timeout_seconds: 600
    max_active_per_session: 1
    disconnected_subscription: execution_continues
    cancellation: cooperative
    process_restart: fail_at_deadline_no_retry
  sse:
    transport: durable_state_and_final_result
    token_deltas: false
    events: [state, final, error, end]
    reconnect: snapshot_same_invocation_no_execution
    resume_event_cursor_supported: true
    current_cursor_range: [1, 3]
paths:
  /api/v1/tenant/agent/sessions:
    get:
      operationId: listServiceSessions
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: agent_session.response.SessionList
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [200]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/SessionList'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
    post:
      operationId: createServiceSession
      x-capability-id: com.corex.agent.session.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateSession'}
      responses:
        '201':
          description: agent_session.response.Session
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [201]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Session'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/tenant/agent/sessions/{session_uuid}:
    get:
      operationId: getServiceSession
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
      responses:
        '200':
          description: agent_session.response.Session
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [200]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Session'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
    patch:
      operationId: renameServiceSession
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RenameSession'}
      responses:
        '200':
          description: agent_session.response.Session
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [200]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Session'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteServiceSession
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
      responses:
        '200':
          description: agent_session.response.Session
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [200]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Session'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/tenant/agent/sessions/{session_uuid}/archive:
    post:
      operationId: archiveServiceSession
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
      responses:
        '200':
          description: agent_session.response.Session
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [200]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Session'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/tenant/agent/sessions/{session_uuid}/messages:
    get:
      operationId: listServiceMessages
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: agent_session.response.MessageList
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [200]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/MessageList'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
    post:
      operationId: appendServiceMessage
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AppendMessage'}
      responses:
        '201':
          description: agent_session.response.Message
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [201]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Message'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/tenant/agent/sessions/{session_uuid}/invocations:
    post:
      operationId: invokeServiceSession
      x-capability-id: com.corex.agent.invoke
      x-additional-capabilities: [com.corex.agent.session.manage]
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Invoke'}
      responses:
        '202':
          description: agent_session.response.Invocation
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [202]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Invocation'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/tenant/agent/sessions/{session_uuid}/invocations/{invocation_uuid}:
    get:
      operationId: getServiceInvocation
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
        - $ref: '#/components/parameters/InvocationUUID'
      responses:
        '200':
          description: agent_session.response.Invocation
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [200]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Invocation'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/tenant/agent/sessions/{session_uuid}/invocations/{invocation_uuid}/cancel:
    post:
      operationId: cancelServiceInvocation
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
        - $ref: '#/components/parameters/InvocationUUID'
      responses:
        '202':
          description: agent_session.response.Invocation
          content:
            application/json:
              schema:
                type: object
                required: [code, data]
                properties:
                  code: {type: integer, enum: [202]}
                  message: {type: string}
                  timestamp: {type: integer, format: int64}
                  request_id: {type: string}
                  data: {$ref: '#/components/schemas/Invocation'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/tenant/agent/sessions/{session_uuid}/invocations/{invocation_uuid}/events:
    get:
      operationId: streamServiceInvocation
      x-capability-id: com.corex.agent.session.manage
      parameters:
        - $ref: '#/components/parameters/SessionUUID'
        - $ref: '#/components/parameters/InvocationUUID'
      responses:
        '200':
          description: agent_session.response.SSE
          content:
            text/event-stream:
              schema: {type: string}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    STS: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    SessionUUID:
      name: session_uuid
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    InvocationUUID:
      name: invocation_uuid
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, maximum: 1000000, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
  responses:
    Error:
      description: agent_session.response.error
      content:
        application/json:
          schema:
            type: object
            required: [code, error_code, reason_code]
            properties:
              code: {type: integer}
              message: {type: string}
              error_code: {$ref: '#/components/schemas/ReasonCode'}
              reason_code: {$ref: '#/components/schemas/ReasonCode'}
  schemas:
    UUID:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
    ReasonCode:
      type: string
      enum: [AGENT_SESSION_INVALID_ARGUMENT, AGENT_SESSION_UNAUTHORIZED, AGENT_SESSION_FORBIDDEN, AGENT_SESSION_NOT_FOUND, AGENT_SESSION_CONFLICT, AGENT_SESSION_IDEMPOTENCY_EXPIRED, AGENT_SESSION_CONTEXT_EXPIRED, AGENT_SESSION_UPSTREAM_DEPENDENCY, AGENT_SESSION_CANCELLED, AGENT_SESSION_EXECUTION_EXPIRED]
    CreateSession:
      type: object
      additionalProperties: false
      required: [agent_uuid]
      properties:
        agent_uuid: {$ref: '#/components/schemas/UUID'}
        title: {type: string, maxLength: 255}
    RenameSession:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: {type: string, maxLength: 255}
    AppendMessage:
      type: object
      additionalProperties: false
      required: [role, content]
      properties:
        role: {type: string, enum: [user]}
        content: {type: string, minLength: 1, maxLength: 65536}
    Invoke:
      type: object
      additionalProperties: false
      required: [message_uuid]
      properties:
        message_uuid: {$ref: '#/components/schemas/UUID'}
    Session:
      type: object
      additionalProperties: false
      required: [session_uuid, agent_uuid, title, status, revision, created_at, updated_at]
      properties:
        session_uuid: {$ref: '#/components/schemas/UUID'}
        agent_uuid: {$ref: '#/components/schemas/UUID'}
        title: {type: string}
        status: {type: string, enum: [active, archived, deleted]}
        revision: {type: integer, format: int64, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Message:
      type: object
      additionalProperties: false
      required: [message_uuid, session_uuid, role, content, sequence, created_at]
      properties:
        message_uuid: {$ref: '#/components/schemas/UUID'}
        session_uuid: {$ref: '#/components/schemas/UUID'}
        role: {type: string, enum: [user, assistant]}
        content: {type: string}
        response_envelope:
          type: object
          description: 已验证的结构化报告；存在时 content 可以为空，客户端应只渲染一份报告。
          additionalProperties: true
        sequence: {type: integer, format: int64, minimum: 1}
        created_at: {type: string, format: date-time}
    Invocation:
      type: object
      additionalProperties: false
      required: [invocation_uuid, session_uuid, message_uuid, status, reason_code, output, trace_uuid, created_at, deadline_at, finished_at]
      properties:
        invocation_uuid: {$ref: '#/components/schemas/UUID'}
        session_uuid: {$ref: '#/components/schemas/UUID'}
        message_uuid: {$ref: '#/components/schemas/UUID'}
        trace_uuid: {$ref: '#/components/schemas/UUID'}
        status: {type: string, enum: [accepted, planning, queued, running, cancelling, succeeded, completed, partial, needs_input, blocked, failed, cancelled]}
        reason_code: {type: string}
        output: {type: string}
        response_envelope:
          type: object
          description: 终态任务回执中的结构化报告，与唯一助手消息保持一致。
          additionalProperties: true
        created_at: {type: string, format: date-time}
        deadline_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time, nullable: true}
    SessionList:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Session'}}
        total: {type: integer, format: int64}
        page: {type: integer}
        page_size: {type: integer}
    MessageList:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Message'}}
        total: {type: integer, format: int64}
        page: {type: integer}
        page_size: {type: integer}